terminado casi ahrex

+++ src/controllers/CategoryController.js
@@ -10,8 +10,8 @@ router.get("/", async (req, res) => {
     const limit = req.query.limit;
     const offset = req.query.offset;
     try {
-      const categorias = await CategoriaService.getCategorias(limit,offset);
-      return res.status(200).json(categorias);
+      const Colection = await CategoriaService.getCategorias(limit,offset);
+      return res.status(200).json(Colection);
     } catch (error) {
       console.log(error);
       return res.json(error);
@@ -42,8 +42,8 @@ router.get("/", async (req, res) => {
 
   router.post("/", AuthMiddleware, async (req, res) => {
     const Categoria = {};
-    Categoria.name = req.query.name;
-    Categoria.display_order = req.query.display_order
+    Categoria.name = req.body.name;
+    Categoria.display_order = req.body.display_order
     
     if (Categoria.name != null) {
       try {
@@ -56,29 +56,36 @@ router.get("/", async (req, res) => {
   } else return res.status(400).json("El nombre (name) esta vacio")
   });
 
-  router.patch("/", authMiddleware, async (req, res) => {
+  router.put("/", authMiddleware, async (req, res) => {
     const categoria = {};
-    categoria.name = req.query.name;
-    categoria.display_order = req.query.display_order;
-    categoria.id = req.query.id;
+    categoria.name = req.body.name;
+    categoria.display_order = req.body.display_order;
+    categoria.id = req.body.id;
 
-    if ((categoria.name != null || categoria.display_order || null) && categoria.id != null) {
+    if ((categoria.name != null || categoria.display_order != null) && categoria.id != null) {
       try {
         const respuesta = await CategoriaService.updateCategoria(categoria)
         return res.status(200).json(respuesta)
       } catch (error) {
         return res.status(400).json(error)
       }
-    } else res.status(404).json("Escribi bien los datos")
+    } else res.status(401).json("bad request")
   })
 
   router.delete("/:id", authMiddleware, async (req,res) => {
     const id = req.params.id;
     try {
-        const respuesta = await CategoriaService.deleteCategoria(id)
+        const cat=await CategoriaService.getCategoriaById(id)
+        if (cat!=null) {
+          const respuesta = await CategoriaService.deleteCategoria(id)
         return res.status(200).json(respuesta)
+        }else{
+          return res.status(404).json("Id no encontrado");
+
+        }
+        
       } catch (error) {
-        return res.status(404).json("Id no encontrado");
+        return res.status(500).json(error)
       }
 
   })
